important post-comp changes
The intake arm now automatically comes up when the intaking command group ends. I managed to do this by manually scheduling the command with the "finallyDo()" decorator.

The intake command group is also set to not be interruptible. This may conflict with the previous statement.

The auto aim is mostly fixed: the issue with it always aiming to the right was a mixture of the limelight not being adjusted and me inputting the speaker position wrong. In addition, the intake pivot was also going too far in and touching the shooter, which impacted the ability for the shooter to pivot.