Saturday changes
Tested RobotGotoFieldPos and it (mostly) works!

Confirmed that the bottom shooter wheels should run faster than the top ones for scoring the amp.

Messed around with the auton some more and found that it specifically errors when I use both a sequential command group and a sendable chooser, even if the sequential command group only has one sendable chooser in it. No clue why this is.