Changes for vision testing
In 2024, we used standard deviations that didn't change, meaning the distance between the robot and the apriltags did not affect how much the pose estimation trusted vision measurements, even though vision measurement were more unstable the farther away the robot was from the apriltags.

For 2025, we want to look into changing the standard deviations depending on how far away the robot is from the apriltags. This code was designed for finding a relationship between robot distance and apriltag consistency, which could then be plugged into a spreadsheet.